back_end/requirements.txt
appdirs==1.4.4
argcomplete==1.12.2
awscli==1.18.211
boto3==1.16.56
botocore==1.19.56
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.4
cfn-flip==1.2.3
chardet==3.0.2
click==7.1.2
colorama==0.4.3
cryptography==3.3.1
distlib==0.3.1
docutils==0.15.2
durationpy==0.5
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Session==0.3.2
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
pip-tools==5.5.0
placebo==0.9.0
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
PyYAML==5.3.1
requests==2.25.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
spotipy==2.16.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.56.0
troposphere==2.6.3
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.4.0
